+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## [Unreleased]
+
+### Added
+
+- 初始化 Flutter 客户端 `qtcloud_pay_studio`（`src/studio`）
+
 ## [0.0.1] - 2026-07-11
 
 阶段晋级至 v0.0.1